% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNearestHigh.R
\name{GetNearestHigh}
\alias{GetNearestHigh}
\title{Determine the time and height of the nearest (by time) high tide for given location.}
\usage{
GetNearestHigh(d, location)
}
\arguments{
\item{d}{a POSIXct value representing the datetime value to be passed to xtide}

\item{location}{string representing the station location as expected by xtide}
}
\value{
a named list of high tide height and time of nearest high tide
}
\description{
This relies on the xtide library to determine the high tide statistics
for the specified time and location.
}
\examples{
#no examples yet
}
