% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNearestLow.R
\name{GetNearestLow}
\alias{GetNearestLow}
\title{Determine the time and height of the nearest (by time) low tide for given location.}
\usage{
GetNearestLow(d, location)
}
\arguments{
\item{d}{a POSIXct value representing the datetime value to be passed to xtide}

\item{location}{string representing the station location as expected by xtide}
}
\value{
a named list of low tide height and time of nearest low tide
}
\description{
This relies on the xtide library to determine the low tide statistics
for the specified time and location.
}
\examples{
#no examples yet
}
