% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTideHeight.R
\name{GetTideHeight}
\alias{GetTideHeight}
\title{Determine the absolute height of the provided datetime for given location.}
\usage{
GetTideHeight(d, location)
}
\arguments{
\item{d}{a POSIXct value representing the datetime value to be passed to xtide}

\item{location}{string representing the station location as expected by xtide}
}
\value{
height in meters
}
\description{
This relies on the xtide library to determine the tide statistics
for the specified time and location.
}
\examples{
#no examples yet
}
